Add files via upload
This was my first work during my 400L IT, done with excel. It's a work to understand if AI and Robotics people, Statistics people, and other Disciplines agree that AI and Robotics significantly affect Statistics. The analysis was done using Weighted Average Rate (WAR) and the overall analysis suggests a strong agreement among the Statistics and AI and Robotics groups that AI and Robotics can effectively solve statistics problems and are seen as an improved method. However, there is fewer consensuses on the ease of use, particularly from the AI and Robotics group. The "Others" group generally shows less strong agreement across all evaluated aspects.